Courses Module Completed And Components Also Updated Auth Modules Also Updated
Lesson Module Remeaning ill Work On it

+++ app/Repositories/Courses/Repository/CourseRepository.php
@@ -7,6 +7,7 @@ use App\Models\Course;
 use App\Models\User;
 use App\Repositories\Courses\Interface\CoursesRepositoryInterface;
 use Cloudinary\Cloudinary;
+use Exception;
 use Illuminate\Http\Request;
 use Illuminate\Support\Str;
 use Spatie\Permission\Models\Role;
@@ -28,11 +29,22 @@ class CourseRepository implements CoursesRepositoryInterface
 
     public function getCourses(Request $request)
     {
+        // dd($request->all());
         $courses = $this->course->query()->latest();
 
         $search = $request->filled('search') ? $request->input('search') : null;
         if ($request->filled('search')) {
             $courses = $courses->where('title', 'like', '%'.$search.'%');
+        }
+
+        $category_id = $request->filled('category_id') ? $request->input('category_id') : '';
+        if ($request->filled('category_id')) {
+            $courses = $courses->where('category_id', $category_id);
+        }
+
+        $instructor_id = $request->filled('instructor_id') ? $request->input('instructor_id') : '';
+        if ($request->filled('instructor_id')) {
+            $courses = $courses->where('instructor_id', $instructor_id);
         }
 
         $courses = $courses->with(['instructor', 'category'])->paginate(10);
@@ -46,7 +58,17 @@ class CourseRepository implements CoursesRepositoryInterface
             return $course;
         });
 
-        return ['courses' => $courses, 'search' => $search];
+        $instructors = $this->user->role('instructor')->get();
+        $categories = $this->category->get();
+
+        return [
+            'courses' => $courses,
+            'search' => $search,
+            'instructors' => $instructors,
+            'categories' => $categories,
+            'instructor_id' => $instructor_id,
+            'category_id' => $category_id,
+        ];
 
     }
 
@@ -58,7 +80,7 @@ class CourseRepository implements CoursesRepositoryInterface
             'short_description' => 'required|string|min:10',
             'description' => 'required|string|min:20',
             'thumbnail' => 'required|image|mimes:jpeg,png,jpg|max:2048',
-            'promo_video' => 'nullable|mimetypes:video/mp4|max:10240000',
+            'promo_video' => 'nullable|mimetypes:video/mp4|max:10485760',
             'instructor_id' => 'required|exists:users,id',
             'category_id' => 'required|exists:categories,id',
             'price' => 'required|numeric|min:0',
@@ -78,44 +100,52 @@ class CourseRepository implements CoursesRepositoryInterface
         ]);
 
         if ($request->hasFile('thumbnail')) {
-            $file = $request->file('thumbnail');
-            $renamedFile = time().uniqid();
-            $movedToCloudaniry = $this->cloudinary->uploadApi()->upload($file->getRealPath(), [
-                'folder' => 'ApexUnified_LMS/Courses/Thumbnails',
-                'public_id' => $renamedFile,
-                'resource_type' => 'image',
-                'transformation' => [
-                    'width' => 1280,
-                    'height' => 720,
-                    'crop' => 'limit',
-                    'quality' => 'auto:best',
-                ],
-            ]);
+            try {
+                $file = $request->file('thumbnail');
+                $renamedFile = time().uniqid();
+                $movedToCloudaniry = $this->cloudinary->uploadApi()->upload($file->getRealPath(), [
+                    'folder' => 'ApexUnified_LMS/Courses/Thumbnails',
+                    'public_id' => $renamedFile,
+                    'resource_type' => 'image',
+                    'transformation' => [
+                        'width' => 1280,
+                        'height' => 720,
+                        'crop' => 'limit',
+                        'quality' => 'auto:best',
+                    ],
+                ]);
 
-            if (! $movedToCloudaniry || ! $movedToCloudaniry['secure_url']) {
-                return ['status' => false, 'message' => 'Something went wrong! While uploading Thumbnail.'];
+                if (! $movedToCloudaniry || ! $movedToCloudaniry['secure_url'] || ! $movedToCloudaniry['public_id']) {
+                    return ['status' => false, 'message' => 'Something went wrong! While uploading Thumbnail.'];
+                }
+
+                $validated_req['thumbnail'] = $movedToCloudaniry['secure_url'];
+                $validated_req['thumbnail_public_id'] = $movedToCloudaniry['public_id'];
+            } catch (Exception $e) {
+                return ['status' => false, 'message' => 'Something went wrong! While uploading Thumbnail. '.$e->getMessage()];
             }
-
-            $validated_req['thumbnail'] = $movedToCloudaniry['secure_url'];
-            $validated_req['thumbnail_public_id'] = $movedToCloudaniry['public_id'];
         }
 
         if ($request->hasFile('promo_video')) {
-            $file = $request->file('promo_video');
-            $renamedFile = time().uniqid();
+            try {
 
-            $movedToCloudaniry = $this->cloudinary->uploadApi()->upload($file->getRealPath(), [
-                'folder' => 'ApexUnified_LMS/Courses/PromoVideos',
-                'public_id' => $renamedFile,
-                'resource_type' => 'video',
-            ]);
+                $file = $request->file('promo_video');
+                $renamedFile = time().uniqid();
 
-            if (! $movedToCloudaniry || ! $movedToCloudaniry['secure_url']) {
-                return ['status' => false, 'message' => 'Something went wrong! While uploading video.'];
+                $movedToCloudaniry = $this->cloudinary->uploadApi()->upload($file->getRealPath(), [
+                    'folder' => 'ApexUnified_LMS/Courses/PromoVideos',
+                    'public_id' => $renamedFile,
+                    'resource_type' => 'video',
+                ]);
+                if (! $movedToCloudaniry || ! $movedToCloudaniry['secure_url'] || ! $movedToCloudaniry['public_id']) {
+                    return ['status' => false, 'message' => 'Something went wrong! While uploading video.'];
+                }
+
+                $validated_req['promo_video'] = $movedToCloudaniry['secure_url'];
+                $validated_req['promo_video_public_id'] = $movedToCloudaniry['public_id'];
+            } catch (Exception $e) {
+                return ['status' => false, 'message' => 'Something went wrong! While uploading Video. '.$e->getMessage()];
             }
-
-            $validated_req['promo_video'] = $movedToCloudaniry['secure_url'];
-            $validated_req['promo_video_public_id'] = $movedToCloudaniry['public_id'];
         }
 
         $validated_req['slug'] = Str::slug($validated_req['title']);
@@ -124,26 +154,33 @@ class CourseRepository implements CoursesRepositoryInterface
         $validated_req['meta_description'] = $validated_req['short_description'];
 
         if ($this->course->create($validated_req)) {
-            return true;
+            return ['status' => true, 'message' => 'Course created successfully!'];
         } else {
-            return false;
+            return ['status' => false, 'message' => 'Course Creation Failed Something went wrong!'];
         }
 
     }
 
     public function getCourse(string $id)
     {
-        return $this->course->find($id);
+        $course = $this->course->find($id);
+
+        if (empty($course)) {
+            return ['status' => false, 'message' => 'Course not found!'];
+        }
+
+        return $course;
     }
 
     public function updateCourse(Request $request, string $id)
     {
+
         $validated_req = $request->validate([
             'title' => 'required|string|min:5|max:150',
             'short_description' => 'required|string|min:10',
             'description' => 'required|string|min:20',
             'thumbnail' => 'nullable|image|mimes:jpeg,png,jpg|max:2048',
-            'promo_video' => 'nullable|mimetypes:video/mp4|max:10240000',
+            'promo_video' => 'nullable|mimetypes:video/mp4|max:10485760',
             'instructor_id' => 'required|exists:users,id',
             'category_id' => 'required|exists:categories,id',
             'price' => 'required|numeric|min:0',
@@ -164,41 +201,43 @@ class CourseRepository implements CoursesRepositoryInterface
 
         $course = $this->getCourse($id);
 
-        if (empty($course)) {
-            return false;
-        }
-
         if (empty($course->thumbnail)) {
             $request->validate(['thumbnail' => 'required']);
         }
 
         if ($request->hasFile('thumbnail')) {
 
-            if (! empty($course->thumbnail) && ! empty($course->thumbnail_public_id)) {
-                $this->cloudinary->uploadApi()->destroy($course->thumbnail_public_id);
+            try {
+                $file = $request->file('thumbnail');
+                $renamedFile = time().uniqid();
+
+                $movedToCloudaniry = $this->cloudinary->uploadApi()->upload($file->getRealPath(), [
+                    'folder' => 'ApexUnified_LMS/Courses/Thumbnails',
+                    'public_id' => $renamedFile,
+                    'resource_type' => 'image',
+                    'transformation' => [
+                        'width' => 1280,
+                        'height' => 720,
+                        'crop' => 'limit',
+                        'quality' => 'auto:best',
+                    ],
+                ]);
+
+                if (! $movedToCloudaniry || ! $movedToCloudaniry['secure_url'] || ! $movedToCloudaniry['public_id']) {
+                    return ['status' => false, 'message' => 'Something went wrong! While uploading Thumbnail.'];
+                }
+
+                if (! empty($course->thumbnail) && ! empty($course->thumbnail_public_id)) {
+                    $this->cloudinary->uploadApi()->destroy($course->thumbnail_public_id);
+                }
+
+                $validated_req['thumbnail'] = $movedToCloudaniry['secure_url'];
+                $validated_req['thumbnail_public_id'] = $movedToCloudaniry['public_id'];
+
+            } catch (Exception $e) {
+                return ['status' => false, 'message' => 'Something went wrong! While uploading Thumbnail. '.$e->getMessage()];
             }
 
-            $file = $request->file('thumbnail');
-            $renamedFile = time().uniqid();
-
-            $movedToCloudaniry = $this->cloudinary->uploadApi()->upload($file->getRealPath(), [
-                'folder' => 'ApexUnified_LMS/Courses/Thumbnails',
-                'public_id' => $renamedFile,
-                'resource_type' => 'image',
-                'transformation' => [
-                    'width' => 1280,
-                    'height' => 720,
-                    'crop' => 'limit',
-                    'quality' => 'auto:best',
-                ],
-            ]);
-
-            if (! $movedToCloudaniry || ! $movedToCloudaniry['secure_url']) {
-                return ['status' => false, 'message' => 'Something went wrong! While uploading Thumbnail.'];
-            }
-
-            $validated_req['thumbnail'] = $movedToCloudaniry['secure_url'];
-            $validated_req['thumbnail_public_id'] = $movedToCloudaniry['public_id'];
         } else {
             $validated_req['thumbnail'] = $course->thumbnail;
             $validated_req['thumbnail_public_id'] = $course->thumbnail_public_id;
@@ -206,27 +245,33 @@ class CourseRepository implements CoursesRepositoryInterface
 
         if ($request->hasFile('promo_video')) {
 
-            if (! empty($course->promo_video) && ! empty($course->promo_video_public_id)) {
-                $this->cloudinary->uploadApi()->destroy($course->promo_video_public_id, [
+            try {
+                $file = $request->file('promo_video');
+                $renamedFile = time().uniqid();
+
+                $movedToCloudaniry = $this->cloudinary->uploadApi()->upload($file->getRealPath(), [
+                    'folder' => 'ApexUnified_LMS/Courses/PromoVideos',
+                    'public_id' => $renamedFile,
                     'resource_type' => 'video',
                 ]);
+
+                if (! $movedToCloudaniry || ! $movedToCloudaniry['secure_url'] || ! $movedToCloudaniry['public_id']) {
+                    return ['status' => false, 'message' => 'Something went wrong! While uploading video.'];
+                }
+
+                if (! empty($course->promo_video) && ! empty($course->promo_video_public_id)) {
+                    $this->cloudinary->uploadApi()->destroy($course->promo_video_public_id, [
+                        'resource_type' => 'video',
+                    ]);
+                }
+
+                $validated_req['promo_video'] = $movedToCloudaniry['secure_url'];
+                $validated_req['promo_video_public_id'] = $movedToCloudaniry['public_id'];
+
+            } catch (Exception $e) {
+                return ['status' => false, 'message' => 'Something went wrong! While uploading video. '.$e->getMessage()];
             }
 
-            $file = $request->file('promo_video');
-            $renamedFile = time().uniqid();
-
-            $movedToCloudaniry = $this->cloudinary->uploadApi()->upload($file->getRealPath(), [
-                'folder' => 'ApexUnified_LMS/Courses/PromoVideos',
-                'public_id' => $renamedFile,
-                'resource_type' => 'video',
-            ]);
-
-            if (! $movedToCloudaniry || ! $movedToCloudaniry['secure_url']) {
-                return ['status' => false, 'message' => 'Something went wrong! While uploading video.'];
-            }
-
-            $validated_req['promo_video'] = $movedToCloudaniry['secure_url'];
-            $validated_req['promo_video_public_id'] = $movedToCloudaniry['public_id'];
         } else {
             $validated_req['promo_video'] = $course->promo_video;
             $validated_req['promo_video_public_id'] = $course->promo_video_public_id;
@@ -244,39 +289,7 @@ class CourseRepository implements CoursesRepositoryInterface
     {
         $course = $this->getCourse($id);
 
-        if (empty($course)) {
-            return false;
-        }
-
-        if (! empty($course->thumbnail) && ! empty($course->thumbnail_public_id)) {
-            $this->cloudinary->uploadApi()->destroy($course->thumbnail_public_id);
-        }
-
-        if (! empty($course->promo_video) && ! empty($course->promo_video_public_id)) {
-            $this->cloudinary->uploadApi()->destroy($course->promo_video_public_id, [
-                'resource_type' => 'video',
-            ]);
-        }
-
-        return $course->delete();
-    }
-
-    public function destroyCourseBySelection(Request $request)
-    {
-        $ids = $request->array('ids');
-        $deleted = 0;
-
-        if (blank($ids)) {
-            return false;
-        }
-
-        $courses = $this->course->whereIn('id', $ids)->get();
-
-        if ($courses->isEmpty()) {
-            return false;
-        }
-
-        foreach ($courses as $course) {
+        try {
             if (! empty($course->thumbnail) && ! empty($course->thumbnail_public_id)) {
                 $this->cloudinary->uploadApi()->destroy($course->thumbnail_public_id);
             }
@@ -287,11 +300,58 @@ class CourseRepository implements CoursesRepositoryInterface
                 ]);
             }
 
-            $course->delete();
-            $deleted++;
+            return $course->delete() ?
+        ['status' => true, 'message' => 'Course deleted successfully!']
+        :
+        ['status' => false, 'message' => 'Something went wrong!'];
+
+        } catch (Exception $e) {
+            return ['status' => false, 'message' => 'Something went wrong! '.$e->getMessage()];
         }
 
-        return $deleted === count($ids);
+    }
+
+    public function destroyCourseBySelection(Request $request)
+    {
+        $ids = $request->array('ids');
+        $deleted = 0;
+
+        if (blank($ids)) {
+            return ['status' => false, 'message' => 'Please select at least one course!'];
+        }
+
+        $courses = $this->course->whereIn('id', $ids)->get();
+
+        if ($courses->isEmpty()) {
+            return ['status' => false, 'message' => 'No Course Found With The Given IDs!'];
+        }
+
+        try {
+
+            foreach ($courses as $course) {
+                if (! empty($course->thumbnail) && ! empty($course->thumbnail_public_id)) {
+                    $this->cloudinary->uploadApi()->destroy($course->thumbnail_public_id);
+                }
+
+                if (! empty($course->promo_video) && ! empty($course->promo_video_public_id)) {
+                    $this->cloudinary->uploadApi()->destroy($course->promo_video_public_id, [
+                        'resource_type' => 'video',
+                    ]);
+                }
+
+                $course->delete();
+                $deleted++;
+            }
+
+            return $deleted === count($ids) ?
+              ['status' => true, 'message' => 'Course deleted successfully!']
+              :
+              ['status' => false, 'message' => 'Something went wrong!'];
+
+        } catch (Exception $e) {
+            return ['status' => false, 'message' => 'Something went wrong! '.$e->getMessage()];
+        }
+
     }
 
     public function getInstructors()
